test(server): centralized update tests
